modify luis file resolver
